GN-117 Refactor Nav component to class. Add custom hook and context for open dropdown state and nav area click outside detection
