+++ src/Header/Nav/Nav.jsx
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React, { useContext } from "react";
 import PropTypes from "prop-types";
 import classnames from "classnames";
 
@@ -11,14 +11,18 @@ import {
 	headerClickEventAction,
 } from "../../tracker";
 
-export default class Nav extends Component {
-	constructor(props) {
-		super(props);
+import useClickOutside from "../../useClickOutside";
+import { GlobalNavContext } from "../../GlobalNavContext";
 
-		this.handleAccountNavItemClick = this.handleAccountNavItemClick.bind(this);
-	}
+function Nav(props) {
+	const { accountsLinks } = props;
+	const context = useContext(GlobalNavContext);
+	const { ref } = useClickOutside(
+		context.idOfOpenDropdown,
+		context.setidOfOpenDropdown
+	);
 
-	handleNavItemClick(e) {
+	function handleNavItemClick(e) {
 		e.preventDefault();
 
 		const href = e.currentTarget.getAttribute("href");
@@ -38,7 +42,7 @@ export default class Nav extends Component {
 		);
 	}
 
-	handleAccountNavItemClick(e) {
+	function handleAccountNavItemClick(e) {
 		const href = e.currentTarget.getAttribute("href");
 
 		let eventLabel;
@@ -65,113 +69,112 @@ export default class Nav extends Component {
 		}
 	}
 
-	render() {
-		const { accountsLinks } = this.props;
+	// Links from NICE Accounts is an object so flatten to a loop for easier traversal
+	const accountsLinksArray =
+		accountsLinks &&
+		Object.keys(accountsLinks).map((text) => ({
+			text: text,
+			href: accountsLinks[text],
+		}));
 
-		// Links from NICE Accounts is an object so flatten to a loop for easier traversal
-		const accountsLinksArray =
-			accountsLinks &&
-			Object.keys(accountsLinks).map((text) => ({
-				text: text,
-				href: accountsLinks[text],
-			}));
-
-		// Would need to polyfill Array.prototype.find to rewrite these loops, whilst we support IE
-		// See https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/find for support
-		let activeService = null;
-		let internalService = false;
-		let servicesToDisplay = services.external; //default to displaying external services.
-		for (let i = 0; i < services.internal.length; i++) {
-			const internalRootLink = services.internal[i];
-			if (this.props.service && internalRootLink.id === this.props.service) {
-				internalService = true;
-				activeService = internalRootLink;
-				servicesToDisplay = [internalRootLink]; //unlike external, internal services dosn't display other internal services.
+	// Would need to polyfill Array.prototype.find to rewrite these loops, whilst we support IE
+	// See https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/find for support
+	let activeService = null;
+	let internalService = false;
+	let servicesToDisplay = services.external; //default to displaying external services.
+	for (let i = 0; i < services.internal.length; i++) {
+		const internalRootLink = services.internal[i];
+		if (props.service && internalRootLink.id === props.service) {
+			internalService = true;
+			activeService = internalRootLink;
+			servicesToDisplay = [internalRootLink]; //unlike external, internal services dosn't display other internal services.
+			break;
+		}
+	}
+	if (!internalService) {
+		for (let i = 0; i < services.external.length; i++) {
+			const externalService = services.external[i];
+			if (props.service && externalService.id === props.service) {
+				activeService = externalService;
 				break;
 			}
 		}
-		if (!internalService) {
-			for (let i = 0; i < services.external.length; i++) {
-				const externalService = services.external[i];
-				if (this.props.service && externalService.id === this.props.service) {
-					activeService = externalService;
-					break;
-				}
-			}
+	}
+	let additionalSubMenuLinks = [];
+	for (let i = 0; i < props.additionalSubMenuItems.length; i++) {
+		const additionalSubMenuItem = props.additionalSubMenuItems[i];
+		if (
+			typeof additionalSubMenuItem !== "undefined" &&
+			additionalSubMenuItem.service === props.service &&
+			Array.isArray(additionalSubMenuItem.links)
+		) {
+			additionalSubMenuLinks = additionalSubMenuItem.links.map((link) => ({
+				text: link.text,
+				href: link.url,
+			}));
 		}
-		let additionalSubMenuLinks = [];
-		for (let i = 0; i < this.props.additionalSubMenuItems.length; i++) {
-			const additionalSubMenuItem = this.props.additionalSubMenuItems[i];
-			if (
-				typeof additionalSubMenuItem !== "undefined" &&
-				additionalSubMenuItem.service === this.props.service &&
-				Array.isArray(additionalSubMenuItem.links)
-			) {
-				additionalSubMenuLinks = additionalSubMenuItem.links.map((link) => ({
-					text: link.text,
-					href: link.url,
-				}));
-			}
-		}
-		const subLinks =
-			activeService &&
-			activeService.links &&
-			activeService.links.concat(additionalSubMenuLinks);
+	}
+	const subLinks =
+		activeService &&
+		activeService.links &&
+		activeService.links.concat(additionalSubMenuLinks);
 
-		return (
-			<div
-				id="header-menu"
-				className={classnames(
-					styles.wrapper,
-					{
-						[styles.wrapperExpanded]: this.props.isExpanded,
-					},
-					{
-						[styles.wrapperWithSubLinks]: subLinks,
-					}
-				)}
-			>
-				<nav className={styles.nav} aria-label="primary navigation">
+	return (
+		<div
+			id="header-menu"
+			className={classnames(
+				styles.wrapper,
+				{
+					[styles.wrapperExpanded]: props.isExpanded,
+				},
+				{
+					[styles.wrapperWithSubLinks]: subLinks,
+				}
+			)}
+			ref={ref}
+		>
+			<nav className={styles.nav} aria-label="primary navigation">
+				<div className={styles.menuWrapper}>
+					<NavLinks
+						handleScrim={props.handleScrim}
+						skipLinkId={props.skipLinkId}
+						servicesToDisplay={servicesToDisplay}
+						currentService={props.service}
+						subLinks={subLinks}
+						onNavigating={props.onNavigating}
+					/>
+				</div>
+			</nav>
+			{accountsLinksArray && (
+				<nav
+					aria-label="My account"
+					className={classnames(styles.nav, styles.myAccount)}
+				>
+					{accountsLinksArray.length > 1 && (
+						<h2 className={styles.myAccountHeading}>My account</h2>
+					)}
 					<div className={styles.menuWrapper}>
-						<NavLinks
-							handleScrim={this.props.handleScrim}
-							skipLinkId={this.props.skipLinkId}
-							servicesToDisplay={servicesToDisplay}
-							currentService={this.props.service}
-							subLinks={subLinks}
-							onNavigating={this.props.onNavigating}
-						/>
+						<ul className={styles.menuList}>
+							{accountsLinksArray.map(({ href, text }) => (
+								<li key={href}>
+									<a
+										href={href}
+										className={styles.link}
+										onClick={handleAccountNavItemClick}
+									>
+										{text}
+									</a>
+								</li>
+							))}
+						</ul>
 					</div>
 				</nav>
-				{accountsLinksArray && (
-					<nav
-						aria-label="My account"
-						className={classnames(styles.nav, styles.myAccount)}
-					>
-						{accountsLinksArray.length > 1 && (
-							<h2 className={styles.myAccountHeading}>My account</h2>
-						)}
-						<div className={styles.menuWrapper}>
-							<ul className={styles.menuList}>
-								{accountsLinksArray.map(({ href, text }) => (
-									<li key={href}>
-										<a
-											href={href}
-											className={styles.link}
-											onClick={this.handleAccountNavItemClick}
-										>
-											{text}
-										</a>
-									</li>
-								))}
-							</ul>
-						</div>
-					</nav>
-				)}
-			</div>
-		);
-	}
+			)}
+		</div>
+	);
 }
+
+export default Nav;
 
 Nav.propTypes = {
 	skipLinkId: PropTypes.string,
