+++ src/Header/Nav/NavLinks/NavLinks.jsx
@@ -7,6 +7,7 @@ import ChevronDown from "@nice-digital/icons/lib/ChevronDown";
 import ChevronUp from "@nice-digital/icons/lib/ChevronUp";
 import SubNav from "../SubNav";
 import Dropdown from "../Dropdown";
+import { GlobalNavContext } from "../../../GlobalNavContext";
 
 import styles from "./NavLinks.module.scss";
 import {
@@ -14,10 +15,6 @@ import {
 	defaultEventCategory,
 	headerClickEventAction,
 } from "../../../tracker";
-import {
-	GlobalNavContext,
-	GlobalNavContextProvider,
-} from "../../../GlobalNavContext";
 
 export function NavLinks({
 	servicesToDisplay,
@@ -27,10 +24,8 @@ export function NavLinks({
 	skipLinkId,
 	handleScrim,
 }) {
-	// const [idOfOpenDropdown, setidOfOpenDropdown] = useState(null);
 	const [canUseDOM, setCanUseDOM] = useState(false);
-
-	const { handleChangeView, idOfOpenDropdown, setidOfOpenDropdown } =
+	const { idOfOpenDropdown, setidOfOpenDropdown } =
 		useContext(GlobalNavContext);
 
 	useEffect(() => {
@@ -41,7 +36,6 @@ export function NavLinks({
 
 	useEffect(() => {
 		handleScrim(Boolean(idOfOpenDropdown !== null));
-		handleChangeView(idOfOpenDropdown);
 	}, [idOfOpenDropdown]);
 
 	function handleNavButtonClick(id) {
@@ -69,24 +63,6 @@ export function NavLinks({
 		if (ESCAPE_KEYS.includes(String(key))) setidOfOpenDropdown(null);
 	}
 
-	// NOTE: could the following be solved with context?
-	function clickOutsideNav(e) {
-		var thingYouClickedOn = e.target;
-		var areaToAvoid = document.getElementById("header-menu");
-		if (!areaToAvoid.contains(thingYouClickedOn)) {
-			console.log("non reacty setting id to null");
-			setidOfOpenDropdown(null);
-		}
-	}
-
-	// useEventListener(
-	// 	"click",
-	// 	clickOutsideNav,
-	// 	// document.querySelector("#global-nav-header")
-	// 	document.querySelector("body")
-	// );
-	// ---------------
-
 	useEventListener("keydown", escapeDropdown);
 
 	const options = {
@@ -95,104 +71,98 @@ export function NavLinks({
 	};
 
 	return (
-		<GlobalNavContextProvider>
-			<p>State from global nav context ### : {idOfOpenDropdown}</p>
-			<FocusTrap active={idOfOpenDropdown !== null} focusTrapOptions={options}>
-				<ul className={styles.menuList} aria-labelledby="header-menu-button">
-					{servicesToDisplay.map(
-						(
-							{
-								href,
-								id,
-								text,
-								abbreviation,
-								title,
-								dropdown,
-								dropdownComponent,
-							},
-							index
-						) => {
-							let ariaCurrent = null;
+		<FocusTrap active={idOfOpenDropdown !== null} focusTrapOptions={options}>
+			<ul className={styles.menuList} aria-labelledby="header-menu-button">
+				{servicesToDisplay.map(
+					(
+						{
+							href,
+							id,
+							text,
+							abbreviation,
+							title,
+							dropdown,
+							dropdownComponent,
+						},
+						index
+					) => {
+						let ariaCurrent = null;
 
-							if (currentService && id === currentService) {
-								ariaCurrent = true;
+						if (currentService && id === currentService) {
+							ariaCurrent = true;
 
-								if (
-									typeof location !== "undefined" &&
-									location &&
-									href ===
-										`${location.protocol}//${location.host}${location.pathname}`
-								) {
-									ariaCurrent = "page";
-								}
+							if (
+								typeof location !== "undefined" &&
+								location &&
+								href ===
+									`${location.protocol}//${location.host}${location.pathname}`
+							) {
+								ariaCurrent = "page";
 							}
+						}
 
-							return (
-								<li key={id} id={id}>
-									{dropdown && canUseDOM ? (
-										<button
-											onClick={() => handleNavButtonClick(id)}
-											aria-current={ariaCurrent}
-											className={styles.link}
-											aria-controls={`dropdown-${id}`}
-											aria-expanded={id === idOfOpenDropdown ? true : false}
-										>
-											<span aria-label={abbreviation && title}>{text}</span>
-											{id === idOfOpenDropdown ? (
-												<ChevronUp
+						return (
+							<li key={id} id={id}>
+								{dropdown && canUseDOM ? (
+									<button
+										onClick={() => handleNavButtonClick(id)}
+										aria-current={ariaCurrent}
+										className={styles.link}
+										aria-controls={`dropdown-${id}`}
+										aria-expanded={id === idOfOpenDropdown ? true : false}
+									>
+										<span aria-label={abbreviation && title}>{text}</span>
+										{id === idOfOpenDropdown ? (
+											<ChevronUp className={styles.icon} pointerEvents="none" />
+										) : (
+											<>
+												<ChevronDown
 													className={styles.icon}
 													pointerEvents="none"
 												/>
-											) : (
-												<>
-													<ChevronDown
-														className={styles.icon}
-														pointerEvents="none"
-													/>
-												</>
-											)}
-										</button>
-									) : (
-										<a
-											href={href}
-											aria-current={ariaCurrent}
-											className={styles.link}
-											onClick={handleNavLinkClick}
-										>
-											<span aria-label={abbreviation && title}>{text}</span>
-										</a>
-									)}
-									{dropdown && canUseDOM ? (
-										<Dropdown
-											component={dropdownComponent}
-											className={classnames([
-												styles.dropdown,
-												id === idOfOpenDropdown && styles.active,
-											])}
-											text={text}
-											nextNavSlug={
-												servicesToDisplay[index + 1]
-													? servicesToDisplay[index + 1]["id"]
-													: skipLinkId
-											}
-											closeDropdown={() => setidOfOpenDropdown(null)}
-											id={`dropdown-${id}`}
-										/>
-									) : null}
-									{ariaCurrent && subLinks && (
-										<SubNav
-											links={subLinks}
-											text={text}
-											onNavigating={onNavigating}
-										/>
-									)}
-								</li>
-							);
-						}
-					)}
-				</ul>
-			</FocusTrap>
-		</GlobalNavContextProvider>
+											</>
+										)}
+									</button>
+								) : (
+									<a
+										href={href}
+										aria-current={ariaCurrent}
+										className={styles.link}
+										onClick={handleNavLinkClick}
+									>
+										<span aria-label={abbreviation && title}>{text}</span>
+									</a>
+								)}
+								{dropdown && canUseDOM ? (
+									<Dropdown
+										component={dropdownComponent}
+										className={classnames([
+											styles.dropdown,
+											id === idOfOpenDropdown && styles.active,
+										])}
+										text={text}
+										nextNavSlug={
+											servicesToDisplay[index + 1]
+												? servicesToDisplay[index + 1]["id"]
+												: skipLinkId
+										}
+										closeDropdown={() => setidOfOpenDropdown(null)}
+										id={`dropdown-${id}`}
+									/>
+								) : null}
+								{ariaCurrent && subLinks && (
+									<SubNav
+										links={subLinks}
+										text={text}
+										onNavigating={onNavigating}
+									/>
+								)}
+							</li>
+						);
+					}
+				)}
+			</ul>
+		</FocusTrap>
 	);
 }
